Add files via upload
Welcome to the Pharaones's after life. Where pharaounes use the Book of the Dead to win the eternal life. 
The texts and images of the Book of the Dead were magical as well as religious. Magic was as legitimate an activity as praying to the gods, even when the magic was aimed at controlling the gods themselves.Indeed, there was little distinction for the Ancient Egyptians between magical and religious practice.

The goal of this game is to win the eternel life on three stage:

Stage 1: 🧟‍ Mummification 🧟‍♂️
Stage 2: 🪦  After life   🪦
Stage 3: ⚖️   Judgment    ⚖️